+++ src/message_dialog.rs
@@ -4,8 +4,7 @@ use std::fmt::{Display, Formatter};
 
 use std::future::Future;
 
-use raw_window_handle::HasWindowHandle;
-use raw_window_handle::RawWindowHandle;
+use raw_window_handle::{HasDisplayHandle, HasWindowHandle, RawDisplayHandle, RawWindowHandle};
 
 /// Synchronous Message Dialog. Supported platforms:
 ///  * Windows
@@ -19,6 +18,7 @@ pub struct MessageDialog {
     pub(crate) level: MessageLevel,
     pub(crate) buttons: MessageButtons,
     pub(crate) parent: Option<RawWindowHandle>,
+    pub(crate) parent_display: Option<RawDisplayHandle>,
 }
 
 // Oh god, I don't like sending RawWindowHandle between threads but here we go anyways...
@@ -66,8 +66,9 @@ impl MessageDialog {
 
     /// Set parent windows explicitly (optional)
     /// Suported in: `macos` and `windows`
-    pub fn set_parent<W: HasWindowHandle>(mut self, parent: &W) -> Self {
+    pub fn set_parent<W: HasWindowHandle + HasDisplayHandle>(mut self, parent: &W) -> Self {
         self.parent = parent.window_handle().ok().map(|x| x.as_raw());
+        self.parent_display = parent.display_handle().ok().map(|x| x.as_raw());
         self
     }
 
@@ -126,7 +127,7 @@ impl AsyncMessageDialog {
 
     /// Set parent windows explicitly (optional)
     /// Suported in: `macos` and `windows`
-    pub fn set_parent<W: HasWindowHandle>(mut self, parent: &W) -> Self {
+    pub fn set_parent<W: HasWindowHandle + HasDisplayHandle>(mut self, parent: &W) -> Self {
         self.0 = self.0.set_parent(parent);
         self
     }